+++ packages/@office-iss/react-native-win32/package.json
@@ -83,7 +83,7 @@
     "@types/prop-types": "15.7.1",
     "@types/react": "^19.0.0",
     "eslint": "^8.19.0",
-    "flow-bin": "^0.261.2",
+    "flow-bin": "^0.267.0",
     "jscodeshift": "^0.14.0",
     "just-scripts": "^1.3.3",
     "prettier": "2.8.8",
