+++ vnext/package.json
@@ -78,7 +78,7 @@
     "@types/node": "^18.0.0",
     "@types/react": "^19.0.0",
     "eslint": "^8.19.0",
-    "flow-bin": "^0.261.2",
+    "flow-bin": "^0.267.0",
     "jscodeshift": "^0.14.0",
     "just-scripts": "^1.3.3",
     "prettier": "2.8.8",
